:art: fix linting
