+++ src/handler/mod.rs
@@ -1,5 +1,6 @@
 #[cfg(feature = "tts")]
-use serenity::model::{channel::Message as SerenityMessage, voice};
+use serenity::model::channel::Message as SerenityMessage;
+use serenity::model::voice;
 use serenity::{
     async_trait,
     client::{Context, EventHandler},
@@ -9,7 +10,6 @@ use serenity::{
 mod model;
 use model::context::Context as Ctx;
 
-#[cfg(feature = "tts")]
 use model::{
     speaker::{CurrentVoiceState, Role},
     voice::Voice,
@@ -18,10 +18,7 @@ pub mod usecase;
 use usecase::set_help_message_to_activity::set_help_message_to_activity;
 
 #[cfg(feature = "tts")]
-use usecase::{
-    speech_welcome_see_you::{leave_if_alone, speech_greeting},
-    text_to_speech::text_to_speech,
-};
+use usecase::{speech_welcome_see_you::speech_greeting, text_to_speech::text_to_speech};
 
 pub struct Handler;
 
@@ -50,6 +47,7 @@ impl EventHandler for Handler {
         new_voice_state: voice::VoiceState,
     ) {
         let state = CurrentVoiceState::new(new_voice_state);
+        #[cfg(feature = "tts")]
         let change = state.change_of_states(old_voice_state.as_ref());
         let member = state.voice_member().await.expect("member is not received");
         let voice = Voice::from(&ctx, member.guild_id).await;
@@ -60,6 +58,12 @@ impl EventHandler for Handler {
         #[cfg(feature = "tts")]
         speech_greeting(&ctx, &voice, &change, &member.user).await;
         leave_if_alone(&ctx, &voice).await;
+    }
+}
+
+async fn leave_if_alone(ctx: &Context, voice: &Voice) {
+    if voice.is_alone(ctx).await.unwrap() {
+        voice.leave().await.unwrap()
     }
 }
