+++ src/handler/model/mod.rs
@@ -1,5 +1,3 @@
 pub mod context;
-#[cfg(feature = "tts")]
 pub mod speaker;
-#[cfg(feature = "tts")]
 pub mod voice;
