+++ src/model/mod.rs
@@ -1,4 +1,5 @@
 mod voice;
 pub use voice::Voice;
 mod message;
+#[cfg(feature = "tts")]
 pub use message::Message;
